Added a bit of documentation + cleanup
